--- Source/Libraries/YpsilonCPU/Emulation/SegmentIndex.cs
@@ -1,11 +1,0 @@
-﻿namespace Ypsilon.Emulation
-{
-    public enum SegmentIndex
-    {
-        CS = 0,
-        DS = 1,
-        ES = 2,
-        SS = 3,
-        IS = 4
-    }
-}
